backend/requirements.txt
boto3==1.24.2          # AWS SDK for Python
flask==2.2.2           # Flask web framework
python-dotenv==0.19.2  # To load environment variables from .env
